requirements.txt
python-dotenv>=1.0.1
pydantic>=2.8.2
rich>=13.7.1
typer>=0.12.3
fastapi>=0.112.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.32
playwright>=1.47.0
